convert to PHP 8 syntax, add typehints, move InvalidOption to Exceptions folder to match structure of other packages
